+++ frontend/components/views/ReportDocView.tsx
@@ -1,8 +1,7 @@
 'use client';
 
 import { useEffect, useRef, useCallback, useMemo, useState } from 'react';
-import { Box, HStack, Text, Button, Input, VStack, Portal, MenuRoot, MenuTrigger, MenuPositioner, MenuContent } from '@chakra-ui/react';
-import { LuChartColumn, LuScanSearch, LuSearch } from 'react-icons/lu';
+import { Box, HStack } from '@chakra-ui/react';
 
 import { LexicalComposer } from '@lexical/react/LexicalComposer';
 import { RichTextPlugin } from '@lexical/react/LexicalRichTextPlugin';
@@ -24,6 +23,7 @@ import { $getRoot, type EditorState } from 'lexical';
 import { editorTheme, EDITOR_NODES, LEXICAL_CONTENT_CSS, ToolbarPlugin } from '../lexical/LexicalTextEditor';
 import { REPORT_TRANSFORMERS } from '../lexical/report-transformers';
 import { QuestionNode, $createQuestionNode } from '../lexical/QuestionNode';
+import ChartPicker from './ChartPicker';
 import { AssetReference } from '@/lib/types';
 import { useAppSelector } from '@/store/hooks';
 
@@ -196,70 +196,13 @@ function ToolbarDivider() {
   return <Box w="1px" h="18px" bg="border.default" mx={1} flexShrink={0} />;
 }
 
-/** Insert-chart picker: lists the pool's questions and inserts the selected one. */
+/** Insert-chart control: inserts the picked question as a chart node at the cursor. */
 function InsertChartControl({ questions }: { questions: ReportQuestion[] }) {
   const [editor] = useLexicalComposerContext();
-  const [search, setSearch] = useState('');
-
-  const filtered = useMemo(() => {
-    const q = search.trim().toLowerCase();
-    return q ? questions.filter(x => x.name.toLowerCase().includes(q)) : questions;
-  }, [questions, search]);
-
   const insert = (questionId: number) => {
     editor.update(() => {
       $insertNodeToNearestRoot($createQuestionNode(questionId));
     });
   };
-
-  return (
-    <MenuRoot positioning={{ placement: 'bottom-end' }} closeOnSelect={false}>
-      <MenuTrigger asChild>
-        <Button size="2xs" variant="subtle" colorPalette="teal" px={2} flexShrink={0} aria-label="Insert chart">
-          <LuChartColumn size={13} />
-          Insert chart
-        </Button>
-      </MenuTrigger>
-      <Portal>
-        <MenuPositioner>
-          <MenuContent minW="280px" bg="bg.surface" borderColor="border.default" shadow="lg" p={2}>
-            <Box position="relative" mb={2}>
-              <Box position="absolute" left={2.5} top="50%" transform="translateY(-50%)" color="fg.muted" pointerEvents="none">
-                <LuSearch size={12} />
-              </Box>
-              <Input
-                placeholder="Search questions…"
-                size="xs"
-                value={search}
-                onChange={(e) => setSearch(e.target.value)}
-                pl={7}
-                autoFocus
-              />
-            </Box>
-            <VStack align="stretch" gap={0} maxH="260px" overflowY="auto">
-              {filtered.length === 0 ? (
-                <Text fontSize="xs" color="fg.muted" px={2} py={3} textAlign="center">No questions</Text>
-              ) : filtered.map(q => (
-                <HStack
-                  key={q.id}
-                  as="button"
-                  aria-label={`Insert ${q.name}`}
-                  gap={2}
-                  px={2}
-                  py={1.5}
-                  borderRadius="sm"
-                  cursor="pointer"
-                  _hover={{ bg: 'bg.muted' }}
-                  onClick={() => insert(q.id)}
-                >
-                  <Box color="accent.primary" flexShrink={0}><LuScanSearch size={13} /></Box>
-                  <Text fontSize="xs" fontWeight={500} color="fg.default" lineClamp={1} textAlign="left">{q.name}</Text>
-                </HStack>
-              ))}
-            </VStack>
-          </MenuContent>
-        </MenuPositioner>
-      </Portal>
-    </MenuRoot>
-  );
+  return <ChartPicker questions={questions} onPick={insert} />;
 }
